Week 2 Challenge/requirements.txt
click>=8.1.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
